Revert to using manual model validation
Reasoning:
1. https://docs.fluentvalidation.net/en/latest/aspnet.html
2. Automatic model validaton was causing all sorts of issues that I was
   unable to resolve in a couple hours. Unexplicable HTTP 500 return
   codes were the most common (string<->enum conversions, missing
   properties or invalid JSON).

For now, model validation is performed by calling the
ValidateToModelStateAsync extension method on the injected validator. If
the validation fails on any of the specified rules, a 400 Bad Request is
returned (actually an instance of ValidationProblem is returned, for
consistent formatting of auto model binding and custom errors).

For the Update endpoints, all fields in the Request should be required.
Unfortunately, when a value type property is missing from the request,
the model binder initializes that property to its default value. A way
around this is to mark the property as nullable, in which case it's
initialized to null. We must then assign a NotNull rule to the property
to mark it as required.

+++ backend/Api/Features/Detectors/Delete.cs
@@ -4,7 +4,6 @@ using System.Threading.Tasks;
 using Api.Domain.Entities;
 using Api.Infrastructure.Database;
 using Ardalis.ApiEndpoints;
-using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Swashbuckle.AspNetCore.Annotations;
@@ -27,28 +26,6 @@ namespace Api.Features.Detectors
             [FromRoute(Name = "id")] public int Id { get; set; }
         }
 
-        public class Validator : AbstractValidator<Req>
-        {
-            private readonly Context _context;
-
-            public Validator(Context context)
-            {
-                _context = context;
-
-                RuleFor(r => r.Id).Must(BeInOffState).WithMessage("Detector must be exist and be in OFF state");
-            }
-
-            private bool BeInOffState(int id)
-            {
-                // TODO(rg): the detector query is duplicated (here and in the action method). Figure out what to do about that
-                var detector = _context.Detectors.SingleOrDefault(d => d.Id == id);
-
-                if (detector is null) return false;
-
-                return detector.State is DetectorState.Off;
-            }
-        }
-
         [HttpDelete(Routes.Detectors.Delete)]
         [SwaggerOperation(
             Summary = "Delete detector",
@@ -59,7 +36,10 @@ namespace Api.Features.Detectors
         public override async Task<ActionResult> HandleAsync([FromRoute] Req req, CancellationToken ct = new())
         {
             var existingDetector = await _context.Detectors.Where(l => l.Id == req.Id).SingleOrDefaultAsync(ct);
-            // NOTE(rg): already checked for null in the validator
+
+            if (existingDetector is null) return NotFound();
+            if (existingDetector.State is not DetectorState.Off)
+                return BadRequest("The Detector must be in the Off state for deletion");
 
             _context.Detectors.Remove(existingDetector);
             await _context.SaveChangesAsync(ct);
